Added algorythm folder
