--- Algorythmn/database/Algorythmn.sql
@@ -1,253 +1,0 @@
--- krijojme databazen
-
-create database algorythmn;
-
--- selektojme databazen ne te cilen duam te punojme
-
-use algorythmn;
--- krijojme tabelen punonjes me fushat perkatese
-
-create table Punonjes(
-id_punonjes int primary key not null,
-emerpunonjesi varchar(45) not null,
-mbiemerpunonjesi varchar(45) not null,
-email varchar(45)  not null
-);
-
-
--- popullojme tablen punonjes
-
-insert into Punonjes values("1","Erlind","Skura","erlindskura@gmail.com"),
-("2","Ilirjan","Kodra","ilirjankodra1@gmail.com"),
-("3","Denis","Laci","lacidenis@gmail.com"),
-("4","Erand","Elbasani","erand.elbassani@outlook.com"),
-("5","Ajrin","Ceku","ajrinceku@gmail.com"),
-("6","Enesa","Karaj","karajenesa@gmail.com"),
-("7","Ana","Fusha","ana_fusha@gmail.com")
-;
-
--- krijojme tablen per prezencen e punonjesve
-
-create table Paraqitja (
-  id_paraqitja int primary key not null,
-  id_punonjes int ,
-  foreign key (id_punonjes) references punonjes(id_punonjes),
-  dita date,
-  prezenca int not null
-);
-
--- popullojme tabelen me vlerat e paraqitjes se punjesve ne zyre per cdo dite te muajit Maj
-
-insert into paraqitja values ("101","1","2023.05.01","0"),
-("102","2","2023.05.01","1"),
-("103","3","2023.05.01","0"),
-("104","4","2023.05.01","1"),
-("105","5","2023.05.01","1"),
-("106","6","2023.05.01","0"),
-("107","7","2023.05.01","1");
-
-
-insert into paraqitja values ("111","1","2023.05.02","1"),
-("112","2","2023.05.02","1"),
-("113","3","2023.05.02","1"),
-("114","4","2023.05.02","1"),
-("115","5","2023.05.02","1"),
-("116","6","2023.05.02","0"),
-("117","7","2023.05.02","1");
-
-insert into paraqitja values ("121","1","2023.05.02","1"),
-("122","2","2023.05.03","0"),
-("123","3","2023.05.03","1"),
-("124","4","2023.05.03","0"),
-("125","5","2023.05.03","1"),
-("126","6","2023.05.03","1"),
-("127","7","2023.05.03","0");
-
-insert into paraqitja values ("131","1","2023.05.04","1"),
-("132","2","2023.05.04","1"),
-("133","3","2023.05.04","1"),
-("134","4","2023.05.04","1"),
-("135","5","2023.05.04","0"),
-("136","6","2023.05.04","1"),
-("137","7","2023.05.04","0");
-
-
-insert into paraqitja values ("141","1","2023.05.05","1"),
-("142","2","2023.05.05","1"),
-("143","3","2023.05.05","0"),
-("144","4","2023.05.05","0"),
-("145","5","2023.05.05","1"),
-("146","6","2023.05.05","1"),
-("147","7","2023.05.05","1");
-
-
-insert into paraqitja values ("151","1","2023.05.08","1"),
-("152","2","2023.05.08","1"),
-("153","3","2023.05.08","1"),
-("154","4","2023.05.08","1"),
-("155","5","2023.05.08","1"),
-("156","6","2023.05.08","0"),
-("157","7","2023.05.08","0");
-
-
-insert into paraqitja values ("161","1","2023.05.09","0"),
-("162","2","2023.05.09","1"),
-("163","3","2023.05.09","1"),
-("164","4","2023.05.09","1"),
-("165","5","2023.05.09","1"),
-("166","6","2023.05.09","1"),
-("167","7","2023.05.09","1");
-
-insert into paraqitja values ("171","1","2023.05.10","1"),
-("172","2","2023.05.10","1"),
-("173","3","2023.05.10","1"),
-("174","4","2023.05.10","1"),
-("175","5","2023.05.10","1"),
-("176","6","2023.05.10","1"),
-("177","7","2023.05.10","1");
-
-insert into paraqitja values ("181","1","2023.05.11","1"),
-("182","2","2023.05.11","0"),
-("183","3","2023.05.11","0"),
-("184","4","2023.05.11","1"),
-("185","5","2023.05.11","1"),
-("186","6","2023.05.11","1"),
-("187","7","2023.05.11","1");
-
-insert into paraqitja values ("191","1","2023.05.02","1"),
-("192","2","2023.05.12","1"),
-("193","3","2023.05.12","1"),
-("194","4","2023.05.12","1"),
-("195","5","2023.05.12","1"),
-("196","6","2023.05.12","1"),
-("197","7","2023.05.12","0");
-
-insert into paraqitja values ("201","1","2023.05.15","1"),
-("202","2","2023.05.15","1"),
-("203","3","2023.05.15","1"),
-("204","4","2023.05.15","1"),
-("205","5","2023.05.15","1"),
-("206","6","2023.05.15","1"),
-("207","7","2023.05.15","0");
-
-insert into paraqitja values ("211","1","2023.05.16","1"),
-("212","2","2023.05.16","1"),
-("213","3","2023.05.16","0"),
-("214","4","2023.05.16","1"),
-("215","5","2023.05.16","0"),
-("216","6","2023.05.16","1"),
-("217","7","2023.05.16","1");
-
-insert into paraqitja values ("221","1","2023.05.17","0"),
-("222","2","2023.05.17","1"),
-("223","3","2023.05.17","1"),
-("224","4","2023.05.17","1"),
-("225","5","2023.05.17","1"),
-("226","6","2023.05.17","0"),
-("227","7","2023.05.17","1");
-
-insert into paraqitja values ("231","1","2023.05.18","1"),
-("232","2","2023.05.18","1"),
-("233","3","2023.05.18","1"),
-("234","4","2023.05.18","1"),
-("235","5","2023.05.18","1"),
-("236","6","2023.05.18","1"),
-("237","7","2023.05.18","1");
-
-insert into paraqitja values ("241","1","2023.05.19","1"),
-("242","2","2023.05.19","0"),
-("243","3","2023.05.19","1"),
-("244","4","2023.05.19","0"),
-("245","5","2023.05.19","1"),
-("246","6","2023.05.19","1"),
-("247","7","2023.05.19","1");
-
-
-insert into paraqitja values ("251","1","2023.05.22","1"),
-("252","2","2023.05.22","1"),
-("253","3","2023.05.22","1"),
-("254","4","2023.05.22","1"),
-("255","5","2023.05.22","0"),
-("256","6","2023.05.22","1"),
-("257","7","2023.05.22","0");
-
-insert into paraqitja values ("261","1","2023.05.23","1"),
-("262","2","2023.05.23","0"),
-("263","3","2023.05.23","1"),
-("264","4","2023.05.23","0"),
-("265","5","2023.05.23","1"),
-("266","6","2023.05.23","1"),
-("267","7","2023.05.23","1");
-
-
-insert into paraqitja values ("271","1","2023.05.24","1"),
-("272","2","2023.05.24","1"),
-("273","3","2023.05.24","1"),
-("274","4","2023.05.24","1"),
-("275","5","2023.05.24","1"),
-("276","6","2023.05.24","1"),
-("277","7","2023.05.24","1");
-
-insert into paraqitja values ("281","1","2023.05.25","0"),
-("282","2","2023.05.25","1"),
-("283","3","2023.05.25","1"),
-("284","4","2023.05.25","1"),
-("285","5","2023.05.25","1"),
-("286","6","2023.05.25","0"),
-("287","7","2023.05.25","1");
-
-insert into paraqitja values ("291","1","2023.05.15","1"),
-("292","2","2023.05.26","1"),
-("293","3","2023.05.26","1"),
-("294","4","2023.05.26","1"),
-("295","5","2023.05.26","0"),
-("296","6","2023.05.26","0"),
-("297","7","2023.05.26","1");
-
-insert into paraqitja values ("301","1","2023.05.29","1"),
-("302","2","2023.05.29","1"),
-("303","3","2023.05.29","1"),
-("304","4","2023.05.29","1"),
-("305","5","2023.05.29","1"),
-("306","6","2023.05.29","1"),
-("307","7","2023.05.29","1");
-
-insert into paraqitja values ("311","1","2023.05.30","1"),
-("312","2","2023.05.30","1"),
-("313","3","2023.05.30","0"),
-("314","4","2023.05.30","1"),
-("315","5","2023.05.30","1"),
-("316","6","2023.05.30","1"),
-("317","7","2023.05.30","1");
-
-insert into paraqitja values ("321","1","2023.05.31","1"),
-("322","2","2023.05.31","1"),
-("323","3","2023.05.31","1"),
-("324","4","2023.05.31","1"),
-("325","5","2023.05.31","1"),
-("326","6","2023.05.31","1"),
-("327","7","2023.05.31","1");
-
-
-select * from paraqitja ;
-
- 
- -- ndertojme querin per fushat qe duam te shfaqim
- 
-select p.id_punonjes, p.emerpunonjesi, p.mbiemerpunonjesi, a.dita, a.prezenca
-from punonjes as p 
-inner join paraqitja as a
-where p.id_punonjes = a.id_punonjes and a.dita="2023.05.16" ;
-
-
-select * from paraqitja;
-
--- query per te afishuar se sa punonjes kane qene ne zyre nga dita e pare e muajit deri ne momentin qe execute prog
-
-select count(a.prezenca) as ditepune, p.id_punonjes, p.emerpunonjesi, p.mbiemerpunonjesi
-from punonjes as p
-inner join paraqitja as a
-on a.id_punonjes = p.id_punonjes
-where  prezenca = 1 and dita between "2023.05.01" and "2023.05.18" 
-group by p.id_punonjes;
-
